reorganized file tree to compliment widget system
