+++ assets/widget-icons/home.png
