+++ components/plugins/loading-bar.js
