+++ index.html
@@ -48,13 +48,13 @@ copyright © 2024-2025
 
 <!--body------------------------>
 <body>
+  <!--body-header----------------------->
     <header>
         <!--header-title------------------------>
         <p>grape.os v0.1.0</p>
         <!--dimmer------------------------>
         <div id="dimmer"></div>
     </header>
-
 
     <!--homepage------------------------>
     <div id="homepage"></div>
@@ -73,7 +73,7 @@ copyright © 2024-2025
     <div id="explorer">
             <div id="explorer-header">
                 <div id="explorer-title"></div>
-                <img src="/assets/icons/close.png" id="close-blog-post">
+                <img src="/assets/graphics/close.png" id="close-blog-post">
             </div>
             <div id="explorer-content"></div>
