+++ widgets/core/notepad/notepad.html
