+++ widgets/core/signup/signup.js
