--- widgets/gabber.html
@@ -1,11 +1,0 @@
-<div class="widget">
-    <div class="widget-title">
-        <h3>Gabber</h3>
-    </div>
-    <div class="widget-content">
-            
-            <div id="gabber-block-01"class="content-block"><p>test</p></div>
-            <div id="gabber-block-02"class="content-block"><p>test 2</p></div>
-    </div>
-
-</div>
