+++ widgets/tertiary/recycle/recycle.html
@@ -6,7 +6,7 @@
         <img src="https://memeshapes.com/cdn/shop/articles/rickastley_x1600.png?v=1712534757" alt="rick-roll" />
        
         <p>oops...</p>
-        <p>probably try not to  <a href="/widgets/oops.html" target="_blank">click here</a>.</p>
+        <p>probably try not to  <a href="/widgets/tertiary/recycle/oops.html" target="_blank">click here</a>.</p>
     </div>
 
 </div>
