+++ widgets/widget-icon-config.js
